+++ apix-impl/src/main/java/eu/xenit/apix/alfresco/properties/PropertyServiceImpl.java
@@ -3,14 +3,17 @@ package eu.xenit.apix.alfresco.properties;
 import com.github.dynamicextensionsalfresco.osgi.OsgiService;
 import eu.xenit.apix.alfresco.ApixToAlfrescoConversion;
 import eu.xenit.apix.dictionary.properties.IPropertyService;
+import eu.xenit.apix.properties.Properties;
 import eu.xenit.apix.properties.PropertyConstraintDefinition;
 import eu.xenit.apix.properties.PropertyFacetable;
 import eu.xenit.apix.properties.PropertyIndexOptions;
 import eu.xenit.apix.properties.PropertyTokenised;
 import java.util.ArrayList;
+import java.util.Collection;
 import java.util.List;
 import org.alfresco.repo.dictionary.Facetable;
 import org.alfresco.repo.dictionary.IndexTokenisationMode;
+import org.alfresco.repo.i18n.MessageService;
 import org.alfresco.service.cmr.dictionary.ConstraintDefinition;
 import org.alfresco.service.cmr.dictionary.DictionaryService;
 import org.alfresco.service.cmr.dictionary.PropertyDefinition;
@@ -28,12 +31,15 @@ public class PropertyServiceImpl implements IPropertyService {
 
     private static final Logger logger = LoggerFactory.getLogger(PropertyServiceImpl.class);
     protected DictionaryService dictionaryService;
+    private MessageService messageService;
     protected ApixToAlfrescoConversion c;
 
     @Autowired
-    public PropertyServiceImpl(DictionaryService dictionaryService, ApixToAlfrescoConversion c) {
+    public PropertyServiceImpl(DictionaryService dictionaryService, ApixToAlfrescoConversion c,
+            MessageService messageService) {
         this.dictionaryService = dictionaryService;
         this.c = c;
+        this.messageService = messageService;
     }
 
     public PropertyIndexOptions GetPropertyIndexOptions(PropertyDefinition definition) {
@@ -48,26 +54,56 @@ public class PropertyServiceImpl implements IPropertyService {
         return IndexOptions;
     }
 
+    private eu.xenit.apix.properties.PropertyDefinition propertyDefinitionConstructor(
+            PropertyDefinition definition) {
+
+        eu.xenit.apix.properties.PropertyDefinition propertyDefinitionUnderConstruction = new eu.xenit.apix.properties.PropertyDefinition(
+                c.apix(definition.getName()), c.apix(definition.getContainerClass().getName()),
+                definition.getTitle(messageService),
+                definition.getDescription(messageService), definition.getDefaultValue(),
+                c.apix(definition.getDataType().getName()), definition.isMultiValued(), definition.isMandatory(),
+                definition.isMandatoryEnforced(), definition.isProtected(), this.GetPropertyIndexOptions(definition),
+                this.GetConstraints(definition));
+
+        return propertyDefinitionUnderConstruction;
+
+    }
+
     public eu.xenit.apix.properties.PropertyDefinition GetPropertyDefinition(eu.xenit.apix.data.QName qname) {
         if (!IsValidPropertyQName(qname)) {
             logger.debug("The given property is no valid property: " + qname.toString());
             return null;
         }
         PropertyDefinition definition = dictionaryService.getProperty(c.alfresco(qname));
-        eu.xenit.apix.properties.PropertyDefinition propertyDefinitionUnderConstruction = new eu.xenit.apix.properties.PropertyDefinition();
-        propertyDefinitionUnderConstruction.setName(c.apix(definition.getName()));
-        propertyDefinitionUnderConstruction.setContainer(c.apix(definition.getContainerClass().getName()));
-        propertyDefinitionUnderConstruction.setTitle(definition.getTitle());
-        propertyDefinitionUnderConstruction.setDescription(definition.getDescription());
-        propertyDefinitionUnderConstruction.setDefaultValue(definition.getDefaultValue());
-        propertyDefinitionUnderConstruction.setDataType(c.apix(definition.getDataType().getName()));
-        propertyDefinitionUnderConstruction.setMultiValued(definition.isMultiValued());
-        propertyDefinitionUnderConstruction.setMandatory(definition.isMandatory());
-        propertyDefinitionUnderConstruction.setEnforced(definition.isMandatoryEnforced());
-        propertyDefinitionUnderConstruction.setIsProtected(definition.isProtected());
-        propertyDefinitionUnderConstruction.setIndexed(this.GetPropertyIndexOptions(definition));
-        propertyDefinitionUnderConstruction.setConstraints(this.GetConstraints(definition));
-        return propertyDefinitionUnderConstruction;
+        return propertyDefinitionConstructor(definition);
+    }
+
+
+    public eu.xenit.apix.properties.PropertyDefinition GetPropertyDefinitionFromAlfrescoQname(
+            org.alfresco.service.namespace.QName qname) {
+
+        PropertyDefinition definition = dictionaryService.getProperty(qname);
+        return propertyDefinitionConstructor(definition);
+
+    }
+
+    @Override
+    public Properties getProperties() {
+
+        Collection<org.alfresco.service.namespace.QName> properties = dictionaryService
+                .getAllProperties(null);
+
+        List<eu.xenit.apix.properties.PropertyDefinition> ret = new ArrayList<>(properties.size());
+        for (org.alfresco.service.namespace.QName property : properties) {
+
+            eu.xenit.apix.properties.PropertyDefinition propertyDefinition = GetPropertyDefinitionFromAlfrescoQname(
+                    property);
+            assert propertyDefinition != null;
+            ret.add(propertyDefinition);
+
+        }
+        return new Properties(ret);
+
     }
 
     public boolean IsValidPropertyQName(eu.xenit.apix.data.QName qname) {